+++ NEWS.md
@@ -1,9 +1,10 @@
 # eemR 0.1.4 (unreleased)
 
+- `eem_raman_normalisation()` now interpolates blank EEM to ensure that em at 350 and excitation between 371 and 428 exist (#31).
+
 - `eem_remove_blank()` and `eem_raman_normalisation()` will now keep blank samples when automatic correction is used. When automatic correction is used, the *untransformed* blank sample will be keep in the list.
 
 - An error will now occur if trying to perform blank correction after Raman normalization. 
- 
 
 # eemR 0.1.3